third_party/WebKit/LayoutTests/fast/dom/shadow/shadow-dom-mouse-event-adjust-offset-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests to ensure that MouseEvent's offsetX and offsetY are adjusted in re-targeting.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS eventRecords['parent']['offsetY'] is 5
PASS eventRecords['shadow-host']['offsetY'] is 5 + 20
PASS eventRecords['top']['offsetY'] is 5 + 20
PASS eventRecords['parent']['offsetX'] is 80 / 2
PASS eventRecords['shadow-host']['offsetX'] is 80 / 2 + 100
PASS eventRecords['top']['offsetX'] is 80 / 2 + 100
PASS successfullyParsed is true

TEST COMPLETE
